+++ apps/server/src/routes/playbackRouter.ts
@@ -11,26 +11,11 @@ import {
   pbStop,
   pbUnload,
 } from '../controllers/playbackController.js';
-import { dispatchFromAdapter } from '../controllers/integrationController.js';
 
 export const router = express.Router();
 
 // create route between controller and '/playback/' endpoint
 router.get('/', pbGet);
-
-// anything else in GET we send to the integration controller
-router.get('/*', (req, res) => {
-  const action = req.path.substring(1);
-  const params = { payload: req.query };
-
-  // TODO: we need to double-check the returns from dispatchFromAdapter
-  try {
-    const reply = dispatchFromAdapter(action, params, 'http');
-    res.status(202).json(reply ?? { message: 'success' });
-  } catch (error) {
-    res.status(500).json({ error: error.message });
-  }
-});
 
 // TODO: we will deprecate this after a while
 // create route between controller and '/playback/start' endpoint
